requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorlog==6.8.2
comm==0.2.2
confection==0.1.5
cymem==2.0.8
datasets==3.0.1
debugpy==1.6.7
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
en_core_web_sm
entrypoints==0.4
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.25.1
hydra-colorlog==1.2.0
hydra-core==1.3.0
idna==3.10
ipykernel==6.29.5
ipython==8.12.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==7.3.4
jupyter_core==5.7.2
langchain==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.2
langcodes==3.4.1
langsmith==0.1.143
language_data==1.2.0
lightning-utilities==0.11.7
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
maverick-coref==1.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
nest_asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.54.4
orjson==3.10.11
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pip==24.2
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.6.0
safetensors==0.4.5
scipy==1.13.1
sentencepiece==0.1.99
sentry-sdk==2.15.0
setproctitle==1.3.3
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
stack-data==0.6.2
sympy==1.13.3
tenacity==9.0.0
thinc==8.2.5
tokenizers==0.20.0
tornado==6.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.7
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wheel==0.44.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.18.3
